FunctionAutomate/requirements.txt
azure-functions
azure-storage-file-share
azure-keyvault-secrets
azure-identity
jinja2
azure-cosmosdb-table
msal